fix: reject an enabled prompt_guard whose backend is not wired up
on_error: block could be silently defeated by config alone. Every field
IsPromptGuardEnabled() requires - jailbreak_mapping_path, and for a remote
protocol an external_models entry with model_role: guardrail carrying
llm_endpoint.address and llm_model_name - was unvalidated. Miss one and that
helper just returns false, which drops jailbreak from the signal dispatch set,
so the guardrail never runs and block becomes a no-op: the exact fail-open it
exists to prevent, with no error anywhere.

Verified by probe on the e2e profile's own config: changing model_role from
guardrail to guard left config.Parse() clean and IsBlock() true while the
signal never dispatched. This also turns the missing-llm_model_name trap found
while bringing up the jailbreak-onerror profile into a config error instead of
a silent startup downgrade.

validator.go's comment claimed these were "checked by
IsPromptGuardEnabled/the main config validation" - no such check existed;
ModelRoleGuardrail appears in pkg/config only as a bool predicate.

Moved the existing backend-selection checks out to validator_prompt_guard.go
alongside the new ones, per pkg/config/AGENTS.md: validator.go is a ratcheted
hotspot and family-specific checks belong in sibling validator files when
extended. validator.go is a net 32 lines smaller.
